use viewset as opposed to view. use separate serializer for userInvite view.

+++ api/views/user_invite.py
@@ -1,30 +1,33 @@
-from rest_framework.views import APIView
+from ..serializers import *
+from ..permissions import IsAuthenticated
+from ..models import TeamInvite, Member, Team
+from rest_framework import serializers
+from rest_framework.response import Response
 from django.db import transaction
+from rest_framework.viewsets import GenericViewSet
+from rest_framework.mixins import ListModelMixin, UpdateModelMixin, DestroyModelMixin
 
-from ..serializers import TeamSerializer
-from .team_related_base import *
 
+class UserInviteViewSet(GenericViewSet,
+                        ListModelMixin,
+                        UpdateModelMixin,
+                        DestroyModelMixin):
 
-class UserInviteView(APIView):
     permission_classes = [IsAuthenticated]
-    serializer_class = TeamSerializer
-    queryset = Team.objects.all().prefetch_related('invite')
+    serializer_class = UserInviteSerializer
+    queryset = TeamInvite.objects.all().prefetch_related('user')
+    pagination_class = None
 
-    def get(self, request) -> Response:
-        teams = TeamSerializer(self.queryset.filter(invite__user=self.request.user), many=True)
-        return Response(teams.data)
+    def update(self, request, *args, **kwargs) -> Response:
+        invite_instance = self.get_object()
+        user_instance = invite_instance.user
+        team_instance = invite_instance.team
 
-    def post(self, request) -> Response:
-        team_name = request.data.get('name', None)
+        if Member.objects.filter(owner=user_instance, team=team_instance).exists():
+            raise serializers.ValidationError({"team": "this member already exists!"})
 
-        team_instance = get_object_or_404(Team, name=team_name)
-        invite_instance = get_object_or_404(TeamInvite, team=team_instance, user=request.user)
-
-        # this should guarantee that if failure occurs, the db won't get into a weird state
         with transaction.atomic():
-            # silently fail for now.
-            member_instance, _ = Member.objects.get_or_create(team=team_instance, owner=request.user)
-            member_instance.save()
+            Member.objects.create(owner=user_instance, team=team_instance).save()
             invite_instance.delete()
 
         return Response({"msg": "success"})
